requirements.txt
Django==1.6.6
django-rest-swagger==0.1.14
djangorestframework==2.3.14
wsgiref==0.1.2
yolk==0.4.3
